move getBackend test helper to common test helpers
